rest_frame/requirements.txt
django~=2.2.4
djangorestframework~=3.11.0